Rename opentelemetry-sdk-extension-aws-v1-support to opentelemetry-sdk-extension-aws.
