--- sdk-extensions/aws-v1-support/README.md
@@ -1,6 +1,0 @@
-# OpenTelemetry AWS Utils
-
-[![Javadocs][javadoc-image]][javadoc-url]
-
-[javadoc-image]: https://www.javadoc.io/badge/io.opentelemetry/opentelemetry-sdk-aws-v1-support.svg
-[javadoc-url]: https://www.javadoc.io/doc/io.opentelemetry/opentelemetry-sdk-aws-v1-support
