--- sdk-extensions/aws-v1-support/src/main/java/io/opentelemetry/sdk/extension/trace/aws/package-info.java
@@ -1,10 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/** Provides implementations of SDK interfaces which integrate natively with AWS infrastructure. */
-@ParametersAreNonnullByDefault
-package io.opentelemetry.sdk.extension.trace.aws;
-
-import javax.annotation.ParametersAreNonnullByDefault;
