--- sdk-extensions/aws-v1-support/src/main/java/io/opentelemetry/sdk/extension/trace/aws/resource/AwsResourceConstants.java
@@ -1,15 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.sdk.extension.trace.aws.resource;
-
-final class AwsResourceConstants {
-
-  static String cloudProvider() {
-    return "aws";
-  }
-
-  private AwsResourceConstants() {}
-}
