--- sdk-extensions/aws-v1-support/src/main/java/io/opentelemetry/sdk/extension/trace/aws/resource/package-info.java
@@ -1,13 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/**
- * {@link io.opentelemetry.sdk.resources.ResourceProvider} implementations for inferring resource
- * information for AWS services.
- */
-@ParametersAreNonnullByDefault
-package io.opentelemetry.sdk.extension.trace.aws.resource;
-
-import javax.annotation.ParametersAreNonnullByDefault;
