--- sdk-extensions/aws-v1-support/src/main/resources/META-INF/services/io.opentelemetry.sdk.resources.ResourceProvider
@@ -1,4 +1,0 @@
-io.opentelemetry.sdk.extension.trace.aws.resource.BeanstalkResource
-io.opentelemetry.sdk.extension.trace.aws.resource.Ec2Resource
-io.opentelemetry.sdk.extension.trace.aws.resource.EcsResource
-io.opentelemetry.sdk.extension.trace.aws.resource.EksResource
