+++ sdk-extensions/aws/README.md
@@ -1,0 +1,6 @@
+# OpenTelemetry AWS Utils
+
+[![Javadocs][javadoc-image]][javadoc-url]
+
+[javadoc-image]: https://www.javadoc.io/badge/io.opentelemetry/opentelemetry-sdk-aws.svg
+[javadoc-url]: https://www.javadoc.io/doc/io.opentelemetry/opentelemetry-sdk-aws
