+++ sdk-extensions/aws/build.gradle
