+++ sdk-extensions/aws/src/main/java/io/opentelemetry/sdk/extension/aws/resource/AwsResourceConstants.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.sdk.extension.aws.resource;
+
+final class AwsResourceConstants {
+
+  static String cloudProvider() {
+    return "aws";
+  }
+
+  private AwsResourceConstants() {}
+}
