+++ sdk-extensions/aws/src/main/java/io/opentelemetry/sdk/extension/aws/resource/package-info.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * {@link io.opentelemetry.sdk.resources.ResourceProvider} implementations for inferring resource
+ * information for AWS services.
+ */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.extension.aws.resource;
+
+import javax.annotation.ParametersAreNonnullByDefault;
