+++ sdk-extensions/aws/src/main/java/io/opentelemetry/sdk/extension/aws/trace/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/** Provides implementations of SDK interfaces which integrate natively with AWS infrastructure. */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.extension.aws.trace;
+
+import javax.annotation.ParametersAreNonnullByDefault;
