+++ sdk-extensions/aws/src/main/resources/META-INF/services/io.opentelemetry.sdk.resources.ResourceProvider
@@ -1,0 +1,4 @@
+io.opentelemetry.sdk.extension.aws.resource.BeanstalkResource
+io.opentelemetry.sdk.extension.aws.resource.Ec2Resource
+io.opentelemetry.sdk.extension.aws.resource.EcsResource
+io.opentelemetry.sdk.extension.aws.resource.EksResource
