+++ settings.gradle
@@ -46,7 +46,7 @@ include ":opentelemetry-all",
         ":opentelemetry-sdk-trace",
         ":opentelemetry-sdk",
         ":opentelemetry-sdk-extension-async-processor",
-        ":opentelemetry-sdk-extension-aws-v1-support",
+        ":opentelemetry-sdk-extension-aws",
         ":opentelemetry-sdk-extension-logging",
         ":opentelemetry-sdk-extension-otproto",
         ":opentelemetry-sdk-extension-resources",
